requirements.txt
streamlit==1.32.0
langchain==0.3.27
python-dotenv
pydantic==2.7.4
groq